+++ src/main/java/au/com/sealink/quicktravel/client/QuickTravelApiClient.java
@@ -11,6 +11,7 @@ import au.com.sealink.quicktravel.client.models.ProductType;
 import au.com.sealink.quicktravel.client.models.Resource;
 import au.com.sealink.quicktravel.client.models.ResourceCategory;
 import au.com.sealink.quicktravel.client.models.Route;
+import au.com.sealink.quicktravel.client.models.Till;
 import au.com.sealink.quicktravel.client.models.User;
 import au.com.sealink.quicktravel.client.models.checkout.Checkout;
 import au.com.sealink.quicktravel.client.models.checkout.CheckoutResponse;
@@ -114,4 +115,6 @@ public interface QuickTravelApiClient {
     @GET("/api/countries.json")
     Single<List<Country>> getCountries();
 
+    @GET("/api/tills.json")
+    Single<List<Till>> getTills();
 }