+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * [TT-4368] Add support for api/routes endpoint
+* [TT-4418] Add support for create/edit scheduled trip endpoints
 
 ## 0.5.0