+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * [TT-4432] Add endpoints for updating and retrieving bookings
 * [TT-4435] Add support for checkout endpoint
 * [TT-4451] Add support for countries endpoint
+* [TT-4452] Add support for tills endpoint
 
 ## 0.5.0