+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * [TT-4435] Add support for checkout endpoint
 * [TT-4451] Add support for countries endpoint
 * [TT-4452] Add support for tills endpoint
+* [TT-4450] Add network interceptor to change tenant hosts
 
 ## 0.5.0